Move generateXML to utils
